refactor: 파일 구조 변경 및 설계 수정
