+++ fe/src/components/IssueList/UIBar/TabButton.tsx
@@ -1,0 +1,5 @@
+function TabButton() {
+	return <div className="component-border">냥냥</div>;
+}
+
+export default TabButton;
