+++ fe/src/components/IssueList/UIBar/UIBar.tsx
@@ -1,9 +1,15 @@
+import Button from "../../common/Button";
 import Filter from "./Filter";
+import TabButton from "./TabButton";
 
 function UI() {
 	return (
-		<div className="container mx-lg my-16 h-[5%] flex">
+		<div className="container mx-lg my-16 flex justify-between h-[40px]">
 			<Filter />
+			<div className="flex h-full w-2/5 bg-sky-100">
+				<TabButton />
+				<Button />
+			</div>
 		</div>
 	);
 }
