+++ fe/src/components/common/DropdownPanel.tsx
@@ -1,10 +1,6 @@
-interface Props {
-	setOpen: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
 const borderBottom = "border-b-2 border-grayscale.300 dark:border-grayscale.600";
 
-function DropdownPanel({ setOpen }: Props) {
+function DropdownPanel() {
 	//TODO : 이하 변수들은 차후 props로 변경 예정
 	const w = "w-[240px]";
 	const filterTitle = "이슈 필터";
